AG.10 clear lint/test blockers and split cross-host modules
